telegramBot/requirements.txt
telebot
python-telegram-bot
requests